pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ghost-ai"
version = "0.2.0"
description = "AI browser agent that costs 50x less. DOM + OCR + Memory. Works with any LLM."
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
keywords = ["ai", "browser", "automation", "agent", "llm", "dom", "ocr"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Libraries",
    "Topic :: Internet :: WWW/HTTP :: Browsers",
]

dependencies = [
    "pillow>=10.0.0",
    "pyautogui>=0.9.54",
    "mss>=9.0.0",
    "openai>=1.50.0",
    "websocket-client>=1.6.0",
    "requests>=2.31.0",
    "rapidocr-onnxruntime>=1.2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.40.0"]
desktop = ["pyobjc-framework-ApplicationServices>=10.0; sys_platform == 'darwin'"]
all = ["ghost-agent[anthropic,desktop]"]

[project.urls]
Homepage = "https://github.com/user/ghost-agent"
Documentation = "https://github.com/user/ghost-agent#readme"
Issues = "https://github.com/user/ghost-agent/issues"

[project.scripts]
ghost = "ghost.core.ghost:Ghost"

[tool.setuptools.packages.find]
include = ["ghost*"]
